debug mode: More reliably find dist
Previously, the dist folder to serve files from in debug mode was found
using the current working directory. This had the disadvantage that it
did not work when the working directory was unexpectedly different.

This commit updates debug mode to more reliably find the dist folder,
using a call to runtime.Caller(0), and assembling an appropriate path.